+++ src/card/cards/combat_constraints.rs
@@ -27,3 +27,6 @@ pub const INSTILL_ENERGY: CardDefinitionId = CardDefinitionId(1826);
 pub const HOWLING_MINE: CardDefinitionId = CardDefinitionId(1827);
 pub const JADE_STATUE: CardDefinitionId = CardDefinitionId(1828);
 pub const CANDELABRA_OF_TAWNOS: CardDefinitionId = CardDefinitionId(1829);
+pub const URZA_S_MINE: CardDefinitionId = CardDefinitionId(1830);
+pub const URZA_S_POWER_PLANT: CardDefinitionId = CardDefinitionId(1831);
+pub const URZA_S_TOWER: CardDefinitionId = CardDefinitionId(1832);