+++ src/game/tests.rs
@@ -649,6 +649,7 @@ mod unleash;
 mod untap_caps;
 mod untap_skips;
 mod until_end_of_combat;
+mod urza_lands;
 mod venarian_gold;
 mod wards;
 mod while_source_tapped;